Extract InnerDeckEditor and InnerDeckStatus wrappers
